+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
   [chris-adam]
 - When a session status is received as completed, mark each signer as signed.
   [sgeulette]
+- Added acroform signature and seal tags validation (PARAF-503).
+  [chris-adam]
 
 1.0b10 (2026-06-18)
 -------------------